# contracts/openapi/mezun360.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "BTÜ Mezun360 API",
    "description": "Session authentication and M2B owner profile, privacy and audited manual verification. Production MFA remains pending.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/me/privacy-preferences": {
      "get": {
        "tags": [
          "privacy-controller"
        ],
        "operationId": "getOwnPrivacy",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {}
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token invalid, CORS rejected or access forbidden.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; sanitized detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "503": {
            "description": "Required dependency unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "privacy-controller"
        ],
        "operationId": "replaceOwnPrivacy",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacyWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {}
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "412": {
            "description": "Exact resource ETag required; stale precondition.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "428": {
            "description": "Exact resource ETag required; stale precondition.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token invalid, CORS rejected or access forbidden.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; sanitized detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "503": {
            "description": "Required dependency unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/admin/verification-requests/{id}/decisions": {
      "post": {
        "tags": [
          "admin-verification-controller"
        ],
        "operationId": "decideVerification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {}
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminVerification"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "412": {
            "description": "Exact resource ETag required; stale precondition.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "428": {
            "description": "Exact resource ETag required; stale precondition.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token invalid, CORS rejected or access forbidden.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; sanitized detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "503": {
            "description": "Required dependency unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/admin/verification-requests": {
      "get": {
        "tags": [
          "admin-verification-controller"
        ],
        "operationId": "listVerifications",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "PENDING",
              "enum": [
                "PENDING",
                "VERIFIED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationQueue"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token invalid, CORS rejected or access forbidden.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; sanitized detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "503": {
            "description": "Required dependency unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/admin/verification-requests/{id}": {
      "get": {
        "tags": [
          "admin-verification-controller"
        ],
        "operationId": "getVerificationReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {}
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminVerification"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid request, missing resource or business conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token invalid, CORS rejected or access forbidden.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; sanitized detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "503": {
            "description": "Required dependency unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PrivacyWrite": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "directoryOptIn": {
            "type": "boolean"
          },
          "profileVisibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "ALUMNI_MEMBERS"
            ]
          }
        },
        "required": [
          "directoryOptIn",
          "profileVisibility"
        ]
      },
      "PrivacyResponse": {
        "type": "object",
        "properties": {
          "profileExists": {
            "type": "boolean"
          },
          "directoryOptIn": {
            "type": "boolean"
          },
          "profileVisibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "ALUMNI_MEMBERS"
            ]
          }
        },
        "required": [
          "directoryOptIn",
          "profileExists",
          "profileVisibility"
        ]
      },
...
      "VerificationDecision": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ]
          },
          "rejectionReason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "status"
        ]
      },
      "AdminVerification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "evidence": {
            "$ref": "#/components/schemas/VerificationEvidence"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ]
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "rejectionReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "current": {
            "type": "boolean"
          }
        },
        "required": [
          "current",
          "evidence",
          "id",
          "rejectionReason",
          "reviewedAt",
          "status",
          "submittedAt"
        ]
      },
      "ClaimedEducation": {
        "type": "object",
        "properties": {
          "institution": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "degree": {
            "type": "string"
          },
          "startYear": {
            "type": "integer",
            "format": "int32"
          },
          "graduationYear": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        },
        "required": [
          "degree",
          "department",
          "graduationYear",
          "institution",
          "startYear"
        ]
      },
      "VerificationEvidence": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "graduationYear": {
            "type": "integer",
            "format": "int32"
          },
          "education": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClaimedEducation"
            }
          }
        },
        "required": [
          "department",
          "education",
          "firstName",
          "graduationYear",
          "lastName"
        ]
      },
      "ApiProblem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "maximum": 599,
            "minimum": 400
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          }
        },
        "required": [
          "code",
          "detail",
          "instance",
          "status",
          "title",
          "traceId",
          "type"
        ]
      },
      "FieldViolation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "field",
          "message"
        ]
      },
...
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "graduationYear": {
            "type": "integer",
            "format": "int32"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ]
          }
        },
        "required": [
          "department",
          "firstName",
          "graduationYear",
          "id",
          "lastName",
          "status",
          "submittedAt"
        ]
      },
      "VerificationQueue": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "items",
          "page",
          "size",
          "totalElements"
        ]
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "mezun360-session",
        "in": "cookie"
      }
    }
  }
}